Refactor check_for_loop_explicit_counter to its own module
